+++ examples/next-2024/next.config.mjs
@@ -1,0 +1,11 @@
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  rewrites: async () => [
+    {
+      source: '/admin',
+      destination: '/admin/index.html',
+    },
+  ],
+}
+
+export default nextConfig