+++ packages/tinacms/src/toolkit/fields/plugins/mdx-field-plugin/index.tsx
@@ -1,20 +1,21 @@
 import React from 'react'
-import { InputProps } from 'react-select/lib/components/Input'
-import { InputFieldType, wrapFieldsWithMeta } from '../wrap-field-with-meta'
-import { RichEditor } from './plate'
+import {
+  type InputFieldType,
+  wrapFieldsWithMeta,
+} from '../wrap-field-with-meta'
 import { EditorContext } from './plate/editor-context'
-import { toolbarItemName } from './plate/plugins/ui/toolbar'
-import type { MdxTemplate } from './plate/types'
 
-// TODO: there's an issue where void node values don't get updated if the editor hasn't been focused from another node first.
-// https://github.com/udecode/plate/issues/1519#issuecomment-1184933602
+import type { MdxTemplate } from './plate/types'
+import { RichEditor } from './plate'
+import type { InputProps } from '@toolkit/fields/components'
+import type { ToolbarOverrideType } from './plate/toolbar/toolbar-overrides'
 
 export type RichTextType = React.PropsWithChildren<
   InputFieldType<
     InputProps,
     {
       templates: MdxTemplate[]
-      toolbarOverride?: toolbarItemName[]
+      toolbarOverride?: ToolbarOverrideType[]
     }
   >
 >
@@ -23,7 +24,7 @@ export const MdxFieldPlugin = {
   name: 'rich-text',
   Component: wrapFieldsWithMeta<
     InputProps,
-    { templates: MdxTemplate[]; toolbarOverride?: toolbarItemName[] }
+    { templates: MdxTemplate[]; toolbarOverride?: ToolbarOverrideType[] }
   >((props) => {
     const [rawMode, setRawMode] = React.useState(false)
     const [key, setKey] = React.useState(0)
@@ -54,7 +55,7 @@ export const MdxFieldPlugin = {
       >
         <div
           className={
-            'min-h-[100px] max-w-full tina-prose relative shadow-inner focus-within:shadow-outline focus-within:border-blue-500 block w-full bg-white border border-gray-200 text-gray-600 focus-within:text-gray-900 rounded-md px-3 py-2'
+            'min-h-[100px] max-w-full tina-prose relative shadow-inner focus-within:shadow-outline focus-within:border-blue-500 block w-full bg-white border border-gray-200 text-gray-600 focus-within:text-gray-900 rounded-md pt-0 py-2'
           }
         >
           {/* {rawMode ? <RawEditor {...props} /> : <RichEditor {...props} />} */}
@@ -76,13 +77,12 @@ export const MdxFieldPluginExtendible = {
       value.children[0].type === 'invalid_markdown'
     ) {
       return 'Unable to parse rich-text'
-    } else {
-      return undefined
     }
+    return undefined
   },
   Component: wrapFieldsWithMeta<
     InputProps,
-    { templates: MdxTemplate[]; toolbarOverride?: toolbarItemName[] }
+    { templates: MdxTemplate[]; toolbarOverride?: ToolbarOverrideType[] }
   >((props) => {
     const [key, setKey] = React.useState(0)
 
@@ -112,7 +112,7 @@ export const MdxFieldPluginExtendible = {
       >
         <div
           className={
-            'min-h-[100px] max-w-full tina-prose relative shadow-inner focus-within:shadow-outline focus-within:border-blue-500 block w-full bg-white border border-gray-200 text-gray-600 focus-within:text-gray-900 rounded-md px-3 py-2'
+            'min-h-[100px] max-w-full tina-prose relative shadow-inner focus-within:shadow-outline focus-within:border-blue-500 block w-full bg-white border border-gray-200 text-gray-600 focus-within:text-gray-900 rounded-md pt-0 py-2'
           }
         >
           {props.rawMode ? props.rawEditor : <RichEditor {...props} />}