+++ playwright/tina-playwright/playwright.config.ts
@@ -32,7 +32,7 @@ export default defineConfig({
 
   /* Run your local dev server before starting the tests */
   webServer: {
-    command: "npm run dev",
+    command: "pnpm dev",
     url: "http://localhost:3000/",
     reuseExistingServer: !process.env.CI,
   },