+++ examples/tina-self-hosted-demo/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @tinacms/self-hosted-starter
 
+## 0.1.57
+
+### Patch Changes
+
+- Updated dependencies [324950a]
+- Updated dependencies [f378f11]
+- Updated dependencies [ceb0c07]
+  - @tinacms/datalayer@1.3.0
+  - tinacms@2.2.0
+  - tinacms-authjs@5.0.0
+  - tinacms-gitprovider-github@2.0.0
+
 ## 0.1.56
 
 ### Patch Changes