+++ packages/@tinacms/vercel-previews/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## 0.0.38
+
+### Patch Changes
+
+- Updated dependencies [324950a]
+- Updated dependencies [f378f11]
+- Updated dependencies [ceb0c07]
+  - tinacms@2.2.0
+
 ## 0.0.37
 
 ### Patch Changes