+++ packages/tinacms-authjs/CHANGELOG.md
@@ -1,5 +1,15 @@
 # tinacms-authjs
 
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies [324950a]
+- Updated dependencies [f378f11]
+- Updated dependencies [ceb0c07]
+  - @tinacms/schema-tools@1.6.0
+  - tinacms@2.2.0
+
 ## 4.0.1
 
 ### Patch Changes