+++ packages/tinacms-clerk/CHANGELOG.md
@@ -1,5 +1,14 @@
 # tinacms-clerk
 
+## 5.0.0
+
+### Patch Changes
+
+- Updated dependencies [324950a]
+- Updated dependencies [f378f11]
+- Updated dependencies [ceb0c07]
+  - tinacms@2.2.0
+
 ## 4.0.1
 
 ### Patch Changes